Move updating buffer counts into buffer usage data module
Note that the tests on the state of the gauges have been dropped as they now
would only test the proper operation of the metrics recorder.
